Completando el CRUD de productos

+++ capstone/mainapp/views.py
@@ -6,9 +6,10 @@ from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.db.models import Q
+from django.urls import reverse
+from django.http import HttpResponseForbidden
 from .models import Producto
 from .forms import RegistrationForm
-
 
 # Create your views here.
 
@@ -25,7 +26,9 @@ def loging_page(request):
         
         if user is not None:
             login(request, user)
-            return redirect('inventory')
+            messages.success(request, 'Login successful. Welcome back!')
+            return redirect('products')
+            
         else:
             messages.error(request, 'Invalid username or password.')
     
@@ -33,14 +36,14 @@ def loging_page(request):
     {
         'title': 'Login'
     })
-
+# vista de contacto
 def contact(request):
     """
     This function handles the contact view.
     """
   
     return render(request,'contact.html')
-
+# vista de about
 def about(request):
     """
     This function handles the about view.
@@ -48,40 +51,56 @@ def about(request):
     return render(request,'about.html')
 
 
-def create_product(request):
-    """
-    This function handles the create_product view.
-    """
-    producto = Producto(
-        title = 'Primer Producto',
-        descripcion = 'Este es el primer producto de la tienda',
-        public = True,
-        
-    )
-    producto.save()
-    return HttpResponse(f'Producto {producto.title} creado exitosamente')
-
 # Editar un producto
-def edit_product(request, id):
-    """
-    This function handles the edit_product view.
-    """
-    producto = Producto.objects.get(pk=id) # Get the product by its id
-    producto.title = 'Producto editado' # Change the title of the product
-    producto.descripcion = 'Este producto ha sido editado' # Change the description of the product
-    producto.save()
-    return HttpResponse(f'Producto {producto.title} editado exitosamente')
+@login_required(login_url='login') # Redirect to the login page if the user is not logged in
+def edit_product(request, id): # request and the product ID
+    if not request.user.is_staff and not request.user.is_superuser:
+        return HttpResponseForbidden("You do not have permission to perform this action")
+    producto = Producto.objects.get(id=id)
+    if request.method == 'POST':
+        producto.title = request.POST.get('title')
+        producto.descripcion = request.POST.get('descripcion')
+        image = request.FILES.get('image') # Get the image from the request
+        if image : 
+            producto.image = image
+        public = 'public' in request.POST # Check if the product is public
+        producto.public = public
+        producto.save()
+        messages.success(request, "Product updated successfully.")
+        return redirect(reverse('products'))
+    return render(request, 'inventory/edit_product.html', {
+        'producto': producto
+    })  
+# Crear Producto
+@login_required(login_url='login') # Redirect to the login page if the user is not logged in
+def create_product(request):
+    if not request.user.is_staff and not request.user.is_superuser:
+        return HttpResponseForbidden("You do not have permission to perform this action")
+    if request.method == 'POST':
+        title = request.POST.get('title')
+        descripcion = request.POST.get('descripcion')
+        image = request.FILES.get('image') # Get the image from the request
+        public = 'public' in request.POST # Check if the product is public
+        producto = Producto.objects.create(title=title, descripcion=descripcion,image=image, public=public)
+        if image: # Check if there is an image in the request
+            producto.image = image
+            producto.save()
+        messages.success(request, "Successfully created product.")
+        return redirect(reverse('products'))
+    return render(request, 'inventory/create_product.html')
 
 # eliminar un producto
-def delete_product(request, id):
-    """
-    This function handles the delete_product view.
-    """
-    producto = Producto.objects.get(pk=id)
+@login_required(login_url='login') # Redirect to the login page if the user is not logged in
+def delete_product(request, id_producto):
+    if not request.user.is_staff and not request.user.is_superuser:
+        return HttpResponseForbidden("You do not have permission to perform this action")
+    producto = Producto.objects.get(id=id_producto)
     producto.delete()
-    return redirect('products')
+    messages.success(request, "Product deleted successfully.")
+    return redirect(reverse('products'))
 
 # listar productos
+@login_required(login_url='login') # Redirect to the login page if the user is not logged in
 def products(request): 
     search_query = request.GET.get('search', '')
     if search_query:
@@ -92,12 +111,12 @@ def products(request):
     return render(request,'inventory/productos.html', {
         'productos': productos
     })
+    
 
-
-def is_superuser_check(user):
+def is_superuser_check(user): # Function to check if a user is a superuser
     return user.is_superuser
 
-@user_passes_test(is_superuser_check)
+@user_passes_test(is_superuser_check) # Only superusers can access this view
 def register_page(request):
     register_form = RegistrationForm()
     
@@ -114,14 +133,15 @@ def register_page(request):
         'title': 'Register',
         'register_form': register_form
     })
-    
-@login_required(login_url='login')
+
+# vista de inventario   
+@login_required(login_url='login') # Redirect to the login page if the user is not logged in
 def inventory_page(request):
     """
     This function handles the inventory view.
     """
     return render(request,'inventory/inventory.html')
-
+# vista de logout
 def logout_user(request):
     logout(request)
     return redirect('login')
